Create README.md
This is a Dungeons and Dragons character stat generator based on the "Roll 4, keep highest 3" rule. That means that you roll 4 dice and you keep the sum of the highest three values as your stat. Do this 6 times... and there you have them! 6 numbers ready to be modified based on your Race and allocated to your Strength, Constitution, Dexterity, Intelligence, Wisdom and Charisma slots as you choose!

To run this generator, simply navigate to the DnD-Character-Stats directory of this repo through a command line window and run:
python DnDCharStats.py